Move EntityVisitor to the right package
