+++ src/main/java/stardeath/visitors/EntityVisitor.java
@@ -1,6 +1,7 @@
-package stardeath.world.visitors;
+package stardeath.visitors;
 
 import stardeath.animates.visitors.AnimateVisitor;
+import stardeath.world.visitors.TileVisitor;
 
 public interface EntityVisitor extends TileVisitor, AnimateVisitor {
 }
