+++ src/main/java/stardeath/visitors/NoOpEntityVisitor.java
@@ -1,4 +1,4 @@
-package stardeath.world.visitors;
+package stardeath.visitors;
 
 public class NoOpEntityVisitor extends DefaultEntityVisitor {
